Refatoraçao completa
